Implementar la lógica de gestión de boletos: se agregaron componentes para la lista de boletos, filtros y pantallas, junto con datos de ejemplo para su visualización.
